requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Werkzeug>=3.0.0
Pillow>=10.0.0
Flask-Migrate==4.1.0
python-dotenv==1.2.2
